+++ PRNcompression/Services/ValidationHelper.cs
@@ -6,11 +6,11 @@ namespace PRNcompression.Services
     {
         /// <summary>byte number string validation</summary>
         /// <returns>byte array size or -1 if validation failed</returns>
-        public static int ValidateNumberString(string byteNumberStr)
+        public static long ValidateNumberString(string byteNumberStr)
         {
             try
             {
-                var x = int.Parse(byteNumberStr);
+                var x = long.Parse(byteNumberStr);
                 if (x < 0) 
                     return -1;
                 return x;
